device/services/vision-capture-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
paho-mqtt==2.1.0
requests==2.32.3
opencv-python-headless==4.10.0.84
numpy==1.26.4
APScheduler==3.10.4
python-dotenv==1.0.1
pydantic==2.8.2
python-multipart==0.0.9
pyserial==3.5

